Refactor excel and pdf files generation
