+++ src/test/java/com/tiagoamp/sjc/model/output/OutputExcelSpreadsheetTest.java
@@ -1,21 +1,16 @@
 package com.tiagoamp.sjc.model.output;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 
-import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
 
 import org.junit.After;
 import org.junit.Before;
-import org.junit.BeforeClass;
 import org.junit.Test;
 
-import com.itextpdf.text.DocumentException;
 import com.tiagoamp.sjc.model.SjcGeneralCode;
 import com.tiagoamp.sjc.model.input.InputExcelSpreadsheet;
 import com.tiagoamp.sjc.model.input.InputSpreadsheet;
@@ -24,12 +19,6 @@ public class OutputExcelSpreadsheetTest {
 	
 	private OutputExcelSpreadsheet excelSpreadsheet;
 
-	@BeforeClass
-	public static void init() throws Exception {
-		Path outDir =  Paths.get("testfiles", "saida");
-		if (Files.notExists(outDir)) Files.createDirectories(outDir);
-	}
-	
 	@Before
 	public void setUp() throws Exception {
 		excelSpreadsheet = new OutputExcelSpreadsheet();
@@ -45,7 +34,7 @@ public class OutputExcelSpreadsheetTest {
 		InputExcelSpreadsheet inputExcelSpreadsheet = new InputExcelSpreadsheet(Paths.get("testfiles", "entrada", "template_input.xlsx"));
 		InputSpreadsheet inputSpreadsheet = null;
 		try {
-			inputSpreadsheet = inputExcelSpreadsheet.toInputSpreadsheet();
+			inputSpreadsheet = inputExcelSpreadsheet.loadFromFile();
 		} catch (IOException e) {
 			e.printStackTrace();
 		}
@@ -53,6 +42,7 @@ public class OutputExcelSpreadsheetTest {
 		list.add(inputSpreadsheet);
 		return list;
 	}
+	
 	
 	@Test
 	public void testLoadDataFromInputSpreadSheets_shouldGenerateOutputSpreadSheet() {
@@ -69,38 +59,7 @@ public class OutputExcelSpreadsheetTest {
 				} else if (sheet.getCode().getGenericCode() == SjcGeneralCode.ADMINISTRATIVO) {
 					assertEquals("'Operacional' sheet should generate 9 rows", 9, sheet.getOutputrows().size());
 				}
-			});
-		
-		/*for (OutSheet sheet : outputSpreadsheet.getSheets()) {
-			if (sheet.getCode().getGenericCode() == SjcGeneralCode.OPERACIONAL) {
-				assertEquals("'Operacional' sheet should generate 17 rows", 17, sheet.getOutputrows().size());
-			} else if (sheet.getCode().getGenericCode() == SjcGeneralCode.ADMINISTRATIVO) {
-				assertEquals("'Operacional' sheet should generate 9 rows", 9, sheet.getOutputrows().size());
-			}
-		}*/
+			});		
 	}
 	
-	@Test
-	public void testGenerateOuputSpreadSheet_shouldGenerateFileInSystem() throws IOException {
-		List<InputSpreadsheet> inputlist = getInputSpreadsheetForTests();
-		excelSpreadsheet.loadDataFromInputSpreadsheets(inputlist);
-		//Path templateFile = Paths.get("resources","template_output.xlsx");
-		Path outputTestFile = Paths.get("testfiles", "saida", "testOutFromSpreadSheetTest.xlsx");
-		
-		excelSpreadsheet.generateOuputSpreadsheetFile(outputTestFile);		
-		//outputSpreadsheet.generateOuputSpreadsheetFile(outputTestFile, templateFile);
-		assertTrue("File should be created in filesystem.", Files.exists(outputTestFile));
-		//assertTrue("Output file must be bigger than original template file.", Files.size(outputTestFile) > Files.size(templateFile) );
-	}
-	
-	@Test
-	public void testGenerateOutputMessageFile_shouldGenerateFileInSystem() throws FileNotFoundException, DocumentException {
-		List<InputSpreadsheet> inputlist = getInputSpreadsheetForTests();
-		excelSpreadsheet.loadDataFromInputSpreadsheets(inputlist);
-		Path outputTestFile = Paths.get("testfiles", "saida", "testOutMessageFromSpreadSheetTest.pdf");
-		
-		excelSpreadsheet.generateOutputMessageFile(outputTestFile);
-		assertTrue("File should be created in filesystem.", Files.exists(outputTestFile));
-	}
-
 }
